requirements-build.txt
# Build-Abhängigkeiten für Nuitka
# Diese Pakete werden nur für den Build-Prozess benötigt

# Nuitka Compiler
nuitka>=1.8.0

# Pillow für Icon-Generierung
Pillow>=10.0.0

# Optional: Ordered Sets für bessere Performance
ordered-set>=4.1.0

# Optional: C Compiler wird benötigt
# Windows: Visual Studio Build Tools oder MinGW64
# Linux: gcc (sudo apt-get install gcc)
# macOS: Xcode Command Line Tools (xcode-select --install)
